Added Bootstrap to Login Page
Some known bugs exist. Couldn't figure out how to vertical alight "Let's
Play" text so it currently has padding-top associated with it. When you
go to a mobile size the blank space on top is greater than the space on
the bottom because of this. Also, when the login fails, the message
causes the whole row to go up. Maybe we want to display the error
differently.